refactor(release) Added api folder for ajax requests feat(fixtures) Added somme upvoters to observes fixtures feat(release) Added json api endpoint for last observes feat(release) Added ajax gif loader
